Ajusted Weapon and Hammer Ability
mod Weapon now makes user that the modifier of the weapon may not put it
in the negatives.

and the hammers passive ability is that when using the useAxe(Weapon)
method it has a chance of hitting the enemies Weapon instead of the
enemy reducing the attack stat of the enemies weapon.